Re-upload of Point of Sale project
A new repository with the Point of Sale application. Version 1 did not upload to github correctly.
This version contains the very basic UI interface, database connectivity and ,most recently, COMM port connectivity, which now allows a connected Cash drawer to be opened at the end of a transaction or when the "Open Register" button is pressed.